+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-## [Unreleased](https://github.com/dotenvx/dotenvx-ruby/compare/v4.0.4...main)
+## [Unreleased](https://github.com/dotenvx/dotenvx-ruby/compare/v4.0.5...main)
+
+## [4.0.5](https://github.com/dotenvx/dotenvx-ruby/compare/v4.0.4...v4.0.5)
+
+- Continue after parse and decryption errors by default, with strict and
+  error-code ignore options for callers that need them.
 
 ## [4.0.4](https://github.com/dotenvx/dotenvx-ruby/compare/v4.0.3...v4.0.4)